+++ toDoList/index.js
@@ -1,7 +1,8 @@
 /** @format */
 
 import {AppRegistry} from 'react-native';
-import App from './App';
+// import App from './App';
+import Header from './android/app/components/Header';
 import {name as appName} from './app.json';
 
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent(appName, () => Header);
